LU-8073 build: Eliminate lustre-source binary package
The lustre source code is already available in tarball form
and on rpm systems in .src.rpm form.  The source code is
also incidentally available in the lustre-debuginfo package.

The reasons for ever having a lustre-source binary package
are probably lost to history.

But this is a really odd thing to do, so we remove that package
from the build system.

Change-Id: I4a175c556d70a380eb550e5ec2830f4a8d3cd04a
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Reviewed-on: http://review.whamcloud.com/19806
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>