LU-8017 obd: report correct health state of a node
The patch 16933 that changed the handling of seq_printf
to work with newer kernels introduced a bug in the proc
file that reports the health of a node. The original
code returned immediately after a healthy state was
reported but the bug prevented the immediate return
so a false unhealthy state was always reported instead.

Change-Id: I414d43d59f0df413ea955acf74c91d2402543228
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: http://review.whamcloud.com/19537
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Frank Zago <fzago@cray.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>