LU-8068 osd-zfs: large dnode support
In ZFS 0.7.x support for large dnode has been added.  This allows
Lustre to optionally specify the size of a dnode which ZFS will
then use to store metadata such as xattrs.

The default dnode size specified by the 'dnodesize' property on the
dataset will be used unless a specific value is provided.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: Iaa7211ca50e7089af4b1fabcccdaa6e16477e641
Reviewed-on: http://review.whamcloud.com/20367
Tested-by: Jenkins
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>