LU-4257 test: Correct error_ignore message
The error_ignore routine requires a bug number to indicate why an
error should be ignored. In sanity test 248, there is no ticket
indicating why the test results should be ignored, but, due to running
the test in a VM, the performance numbers may not be what is
expected and, thus, the results can be ignored.

Add a reason to ignore the error to the call to error_ignore.

Test-Parameters: trivial testlist=sanity,sanity,sanity,sanity

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ieda9d06facd3e1519ac6a84e15505bc4c74aa60a
Reviewed-on: http://review.whamcloud.com/20647
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>