LU-8216 ldiskfs: fix journal quota files for RHEL6.8
When we hit Lustre crash, e2fsck always complains of Quota
accounting mismatch, and sometimes differences are huge.

The problem is that we don't journal quota updates properly, and
with current codes quota files will be only journaled when syncing
the filesystem, which only happens by sync() call or unmount.

RHEL6.8 version of the patch http://review.whamcloud.com/20503

Signed-off-by: Wang Shilong <wshilong@ddn.com>
Change-Id: Id5b580083197555029442338f7feaf59b1ad154c
Reviewed-on: http://review.whamcloud.com/20791
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>