LU-8832 tests: Generalize check of virtual machine
The patch for LU-8248, commit 39582273a7a52fb, we added a
function, called running_in_vm(), to check if the test is
being run in a VM and, if so, ignore the performance
results. The check for running in a VM needs to be updated
to run correctly in a SLES environment.

Test-Parameters: trivial clientdistro=sles11sp4 mdsdistro=sles11sp4 ossdistro=sles11sp4 testlist=sanity
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Iebca20c6be81da7fef18e7d486cb59e3f3877eaa
Reviewed-on: https://review.whamcloud.com/23774
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>