LUDOC-300 misc: cleanup references and obsolete text
Remove references to specific Lustre release versions and supported
Linux distros, since these are very version dependent and are better
described in the release notes for each version.

Fix up packages needed on the client and server.

Remove obsolete LNDs from the supported list.

Fix table and figure references to be on the table/figure itself
and not on the title.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I0bf3788f0ee8c32021b61386c6b69e0c47e64ade
Reviewed-on: https://review.whamcloud.com/23944
Tested-by: Jenkins
Reviewed-by: James Nunez <james.a.nunez@intel.com>
Reviewed-by: Joseph Gmitter <joseph.gmitter@intel.com>