LU-8066 obd: Add debugfs root
This is just plumbing for migrating remaining procfs to
debugfs support

Linux-commit: 5c8c82f63a11c07a0687d2c71411166017012689

ldebugfs_remove is usually called on directories with files passed in
as attributes, so simple debugfs_remove failes on them as not empty
Switch to debugfs_remove_recursive.

This fixes a number of problems where a new filesystem is mounted after
being unmounted first.

Linux-commit: 6a491f2b80f2806221ba3a5a3e26fbe945f82d83

Turns out we mistakenly export some pretty-wide-reaching debugfs
functions as EXPORT_SYMBOL instead of EXPORT_SYMBOL_GPL as we should,
so this patch rectifies the situation.

Linux-commit: da33f1dd9f69c5a5d23e1254906b027af3362e36

Move /proc/fs/lustre/devices to debugfs. The devices file prints out
status information about all obd devices in the system in human
readable form.

Linux-commit: 4361a048743f900bb0710bd7cb36a650d7bef93a

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Change-Id: Iefe6ee69b607f627c5682513128706bf0c1c3e09
Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/23427
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Emoly Liu <emoly.liu@intel.com>