LU-5170 utils: Add support for --list-commands option
A --list-commands option has been added to lfs, lctl, lnetctl and
lst to output a list of the commands supported by each utility. The
commands are printed in a multi-column format to produce more compact
output that is easier to read.  The appropriate man pages have been
updated to include the new options.

The obsolete ACL and join commands have been removed from lfs.

Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Change-Id: I3715049539c76e0cd03accfccfbf7eda6f4bf2ff
Reviewed-on: https://review.whamcloud.com/24902
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Lai Siyao <lai.siyao@intel.com>