LU-9069 tests: improve output of sanity test_255a
Improve output of sanity.sh test_255a to contain more information.
Clean up the performance measurements and calculations to make the
test easier to read.  The random_read_iops() helper routine might
be useful for other tests in the future as well.

The test does not (yet) work for ZFS, so it will skip the checks on
ZFS OSTs until "ladvise -a dontneed" is implemented for osd-zfs.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ia177d87e41266b058a6863bbf36108ad71ef9a00
Reviewed-on: https://review.whamcloud.com/24907
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>