LU-5969 lustreapi: allow "version" without "lustre:"
The upstream kernel /sys/fs/lustre/version file does not have the
"lustre:" prefix in it, only the version string.  Since we no longer
have multiple version strings in this file, if "lustre:" is not
found then just assume the whole string is the version, otherwise
continue to strip out the "lustre:" prefix as we did before.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ife50f83ad422a8e3c9ae857f519a5cf4ca3ebbe5
Reviewed-on: https://review.whamcloud.com/25324
Tested-by: Jenkins
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>