LU-8582 tests: skip sanity test_255a for interop
Move the sanity.sh test_255a version check to the start of the test
instead of after some of the ladvise functionality is tested.

That said, there is still an issue here, because the client shouldn't
hang indefinitely if the server is returning -EOPNOTSUPP, in the case
where a new client/lfs is connecting to an old OSS, and that issue is
not addressed by this patch.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: If1fbc13bd60cfc1942a43aff40e3e7905cdcab07
Reviewed-on: https://review.whamcloud.com/25362
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>