LU-9125 test: Update setstripe options
Some flags for 'lfs setstripe'  will be depricated in
tag 2.9.59; '--count' replaced by --stripe-count or -c.

replay-single test 68 will silently fail due to this change.
We need to check that an error is called if 'lfs setstripe'
fails and change the depricated parameters used in replay-single.
The check_default_stripe_attr() routine in sanity.sh also needs
to be updated with the new setstripe options.

Test-Parameters: trivial testlist=sanity,replay-single

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ie5809c9268684675585d17cd1c402ec3fb002239
Reviewed-on: https://review.whamcloud.com/25475
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Wei Liu <wei3.liu@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>