LUDOC-11 utils: improve mount.lustre.8 option description
Add the "mgssec=<flavor>" mount option for SSK and Kerberos.
Add the "skpath=pathname" mount option for SSK.
Add the "lazystatfs" and "nolazystatfs" mount options.
Add a description of mount-by-label and mount-by-UUID, with caveats.
Improve the description of the "mgsnode" parameter.
Improve the description of the "flock" and "noflock" options.
Move the "acl" and "noacl" mount options from the client to the
server section, since they've been long deprecated on the client.
Fix whitespace.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I2343df49270ff1f18b7bed20042eafff893ebbe5
Reviewed-on: https://review.whamcloud.com/25680
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <joseph.gmitter@intel.com>