LUDOC-11 config: improve ZFS MDT space calculations
Improve the description of inode space consumption for ZFS MDTs,
as well as ZFS reserved space.

Add a note that "lfs df -i" on a newly formatted MDT or OST will
report a very low estimate for the total number of inodes in the
filesystem until some regular files have been created on the MDT.

Remove a duplicate note about Lustre 2.4 being able to expand the
number of inodes by adding DNE MDTs to an existing filesystem.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ic035e1a44b4db35dfa1bf9d7b2ed1b2fb0434739
Reviewed-on: https://review.whamcloud.com/25713
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <joseph.gmitter@intel.com>