LU-9318 test: sanity-lsnapshot skipped for all non-zfs configs
The problem with the script was that it tried to check
for ZFS file system after the script tried to add snapshot
to the debug list.

As part of solution the script just needs to check for ZFS
file system or the version is too old at the start for of
the script.

Test-Parameters: trivial testlist=sanity-lsnapshot,sanity-lsnapshot,sanity-lsnapshot
Signed-off-by: Saurabh Tandan <saurabh.tandan@intel.com>
Change-Id: Ie59dac7543d1e35b924c84d19c34a5ec5b903a14
Reviewed-on: https://review.whamcloud.com/26600
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Nunez <james.a.nunez@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>