LU-9019 libcfs: migrate watchdog to 64 bit time
The watchdog computes timestamps from timeval, which overflows
in 2038 on 32-bit systems. This changes the results output to
timespec64 type to avoid the overflow. The lcw_last_touched
field is changed to ktime_t since better than seconds resolution
is wanted. For lcw_last_watchdog it is changed to time64_t since
we only care about seconds percision. Both changes will avoid
the 2038 overflow issue and the HZ variablity across platforms
for jiffies.

Change-Id: If81f60a408db956540c563fc695f729cd67cdd9e
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/26994
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Olaf Weber <olaf.weber@hpe.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>