LU-9096 test: add sanity 253 to ALWAYS_EXCEPT
Discussion in the LU-9096 ticket questions if sanity
test 253 is a valid/correct test. Since tes 253 is failing
intermittently for ZFS servers add the test to the
ALWAYS_EXCEPT list until the test can be improved.

Since the correctness of the test is being questioned,
skip the test for all server file systems, not just ZFS.

Test-Parameters: trivial
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I14cffb288b4d338ff009f4b57a19d5a5f4c516ff
Reviewed-on: https://review.whamcloud.com/27013
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Wei Liu <wei3.liu@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>