LUDOC-376 build: add condition 'l2A' for 2.10
Add support for annotating content with "condition='l2A'" to indicate
new content for Lustre 2.10.  Also add "condition='l2B'" for 2.11 in
advance.

Remove support for "condition='l21'" which was never used in the
manual. Remove markers for "condition='l22'" which is now very old,
and fix the content to indicate this is "normal" behaviour.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I1daa97cbfe58247cda912b300f1cea31c1dc2e7c
Reviewed-on: https://review.whamcloud.com/27055
Reviewed-by: Joseph Gmitter <joseph.gmitter@intel.com>
Tested-by: Jenkins