LU-9522 test: sanity 27z fix update local variable
sanity test_27z is failing for servers running ZFS. The
problem is that filter_fid is not being updated for each
stripe of the file being reviewed in check_seq_oid().

For each stripe, the variable that holds the filter_fid,
ff, needs to be reinitialized.

Test-Parameters: trivial testgroup=review-zfs-part-1

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I550720a0a2bad22c4ffa98ca84b5f085eb220051
Reviewed-on: https://review.whamcloud.com/27163
Tested-by: Jenkins
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>