LU-6210 utils: Use C99 struct initializers in lustre_fsck.c
This patch makes no functional changes.  The long_opt_stop and
long_opt_query struct initializers in lustre_lfsck.c are updated
to C99 syntax.

C89 positional initializers require values to be placed in the
correct order. This will cause errors if the fields of the struct
definition are reordered or fields are added or removed. C99 named
initializers avoid this problem, and also automatically clear any
values that are not explicitly set.

Test-Parameters: trivial
Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Change-Id: Ie713904f31208447e4d5741023356aa1a37c0ea3
Reviewed-on: https://review.whamcloud.com/27790
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Fan Yong <fan.yong@intel.com>