LU-6210 utils: Use C99 struct initializer in mkfs_lustre.c
This patch makes no functional changes.  The long_opt_start
struct initializer in mkfs_lustre.c is updated to C99 syntax.
The option struct and option string have been renamed to long_opts
and short_opts for consistency.

C89 positional initializers require values to be placed in the
correct order. This will cause errors if the fields of the struct
definition are reordered or fields are added or removed. C99 named
initializers avoid this problem, and also automatically clear any
values that are not explicitly set.

Test-Parameters: trivial
Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Change-Id: Icbc6be0eeaf6bf369a1b593917115378b2e3e159
Reviewed-on: https://review.whamcloud.com/27815
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>