LU-6210 utils: Use C99 struct initializer in parse_opts()
This patch makes no functional changes.  The long_opt initializer in
parse_opts() is updated to C99 syntax, and renamed to long_opts for
consistency.

C89 positional initializers require values to be placed in the
correct order. This will cause errors if the fields of the struct
definition are reordered or fields are added or removed. C99 named
initializers avoid this problem, and also automatically clear any
values that are not explicitly set.

Test-Parameters: trivial
Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Change-Id: Id3f1540c33cbd5db33a4a0eb69a89672eac7f713
Reviewed-on: https://review.whamcloud.com/27817
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>