LU-6210 utils: Use C99 initializer in jt_nodemap_del_range()
This patch makes no functional changes.  The long_options struct
initializer in jt_nodemap_del_range() is updated to C99 syntax.
It is renamed to long_opts for consistency.

C89 positional initializers require values to be placed in the
correct order. This will cause errors if the fields of the struct
definition are reordered or fields are added or removed. C99 named
initializers avoid this problem, and also automatically clear any
values that are not explicitly set.

Test-Parameters: trivial
Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Change-Id: I442d2cb65c1ffefc2229ead83e04b7ec2d8160e5
Reviewed-on: https://review.whamcloud.com/27883
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>