LU-9295 test: skip sanity-lfsck 9 less than 2.8
sanity-lfsck tests 9a and 9b check the version of the MDS
to see if the test should be run. Unfortunately, the
server_version variable is never initialized and the tests
are skipped.

For sanity-lfsck tests 9a and 9b, get the server version
using the lustre_version_code routine.

Test-Parameters: trivial testlist=sanity-lfsck
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I6c461a6940a0cbd0439c8f29d979ef799ea30229
Reviewed-on: https://review.whamcloud.com/26403
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>